[wasm] Add debug output to consumeNBytes (#622)
A person debugging the parser likely wants to be able to easily see a
record of how the thing failed if things went poorly.

Since we now have debug output hidden under a flag, it's probably not
too verbose to output details like "number of bytes consumed" and
"number of bytes remaining."

This might save someone a few minutes futzing around in a debugger.

Since this only impacts debug output, which is allowed to be unstable
(until it's used in a test or something one day), no testcase.

Co-authored-by: Jessica Paquette <jessica.paquette@woven.toyota>