[wasm] Translate + convert rotr (#649)
This adds translation and conversion to arith for

* i32.rotr
* i64.rotr

This uses the same assembly format as shifts. Update Wasm_ShiftOp's name
to reflect that.

Since arith doesn't contain native rotates at this time, for now, let's
translate it to a series of bit twiddling ops.

This contains the pattern suggested by Fabien in this thread:
https://groups.google.com/g/llvm-dev/c/QFPxqJzQJAo

Rather than the popular pattern used in books like Hacker's Delight, in
order to avoid creating a shift with UB when rotating by 0 bits. Since
rotates are fully defined in wasm, we need to preserve that behaviour
when we convert to other dialects.

Annoyingly, LLVM is pretty bad at matching shift patterns, so we might
not end up getting a native rotate out of this expression. I can't find
any particular reason for arith to not support rotates in any existing
discussion. But that's probably something to talk about upstream.

The pattern matching keeps this in the wasm dialect rather than
translating directly to arith. Then, the broken down rotate will be
subsequently translated to arith using our existing conversion.

Co-authored-by: Jessica Paquette <jessica.paquette@woven.toyota>