[wasm] Convert + translate rotl (#651)
Same as rotr, but swap the resultant shifts.

Repurpose the op conversion class so it can be used for both rotates.
It's very simple, only using template parameters, but since there are
only two of these instructions, and this isn't really going to be used
elsewhere, it seemed overkill to do anything more intelligent.

This gives us

* i32.rotl
* i64.rotl

Co-authored-by: Jessica Paquette <jessica.paquette@woven.toyota>