[wasm] Make binary op translation handle any number of opcodes (#654)
We need to support unary ops too.

Make it so that the existing code does not assume we are building a
binary op, and allow any number of operands.

Rename code as required to make it clear that the current code only
handles binary operations.

While we're here, add some debug output to the operation conversion code
to make it clearer what's going on under --debug.

Co-authored-by: Jessica Paquette <jessica.paquette@woven.toyota>