[mlir][wasm] Add test for invalid function type index (#22)
* [wasm] Add debug output to consumeNBytes (#628)

A person debugging the parser likely wants to be able to easily see a
record of how the thing failed if things went poorly.

Since we now have debug output hidden under a flag, it's probably not
too verbose to output details like "number of bytes consumed" and
"number of bytes remaining."

This might save someone a few minutes futzing around in a debugger.

Co-authored-by: Jessica Paquette <jessica.paquette@woven.toyota>

* [mlir][wasm] Introduce wasabi binary driver for wasm to llvm ir translation

Wasabi is a basic driver currently able to take WASM binaries or WASM MLIR
files as input and convert them to standard mlir, llvm mlir or llvm ir.
It doesn't handle the backend part of the conversions.

* [mlir][wasm] Add test for invalid function type index

Verify that the parser will fail on a nonexistent type index.

---------

Co-authored-by: Jessica Paquette (Woven by Toyota, Inc.)/Jessica Paquette <jessica.paquette@woven-planet.global>
Co-authored-by: Ferdinand Lemaire <ferdinand.lemaire@woven-planet.global>